Обновил структуру модуля и вынес настройки
