+++ locale/ru/module-economy.cfg
@@ -2,3 +2,15 @@
 
 enabled-name=Виртуальная экономика
 enabled-desc=Торговля между командами
+salary-interval-name=Интервал выдачи зарплаты (минуты)
+salary-interval-desc=Время между ЗП
+fluid-coefficient-name=Коээфицент выплат для жидкостей
+fluid-coefficient-desc=
+price-ore-name=Изначальная цена руды
+price-ore-desc=
+price-production-name=Наценка за производство
+price-production-desc=
+
+[entity-name]
+
+shop-buy=Магазин
